backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
groq
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
supabase
